[MOL-12849] Create specific types file
